LLM/sql-agent/pyproject.toml
[project]
name = "sql-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.26",
    "langchain[google-genai]>=0.3.26",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.1",
]
